src/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
openai==1.58.1
PyPDF2==3.0.1
elevenlabs==1.9.0
httpx==0.27.2